% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BrcParcellation.R
\name{plot.BrcParcellation}
\alias{plot.BrcParcellation}
\title{Plot method for objects of class `"BrcParcellation"`}
\usage{
\method{plot}{BrcParcellation}(x, numSlices, view = "sagittal",
  colors = NULL, ...)
}
\arguments{
\item{x}{an object of class `"BrcParcellation"`}

\item{numSlices}{the number of slices to plot. Values greater than or equal
to the maximum number of slices will plot all the slices.}

\item{view}{one of three strings: "sagittal", "coronal", or "axial".
Specifies whether to plot sagittal, coronal, or axial slices.}

\item{colors}{A list of R colors similar to that generated by
grDevices::rainbow().  Must contain exactly as many colors as there are
unique parcels. If NULL, a default coloring is used.}

\item{...}{not used}
}
\description{
Plots an evenly-spaced series of slices through the brain. Each parcel is
colored differently. Note that empty slices are trimmed before plotting.
}

