% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.BrcParcellation.R
\name{plot3d}
\alias{plot3d}
\title{Generic 3D Plotting}
\usage{
plot3d(x, ...)
}
\arguments{
\item{x}{the S3 object to plot}

\item{...}{Arguments to be passed to methods, such as graphical parameters.}
}
\description{
Generic function for 3D plotting of R objects.
}
\details{
For detailed documentation, use `methods(plot3d)` and
see the man pages for the individual methods.
}

