% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.BrcFmri.R
\name{plotTimeSeries.BrcFmri}
\alias{plotTimeSeries.BrcFmri}
\title{Plotting Time Series for a BrcFmri Object}
\usage{
\method{plotTimeSeries}{BrcFmri}(x, buffer = NULL, same.screen = TRUE,
  xregion = NULL, yregion = NULL, zregion = NULL, ...)
}
\arguments{
\item{x}{the BrcFmri object}

\item{buffer}{a numeric determine the spacing between two time series in the plot. 
Only used if same.screen = TRUE.}

\item{same.screen}{a boolean to determine if all the time series are put on one plot
or seperate plots}

\item{xregion}{the direction along the first dimension of the BrcFmri data in 3-dimensions
to dictate which voxels have their time series shown}

\item{yregion}{the direction along the second dimension of the BrcFmri data in 3-dimensions
to dictate which voxels have their time series shown}

\item{zregion}{the direction along the third dimension of the BrcFmri data in 3-dimensions
to dictate which voxels have their time series shown}

\item{...}{Arguments to be passed to plotting, such as graphical parameters.}
}
\description{
Creates plots of the time series at various voxels inside the BrcFmri object.
}

