% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTimeSeries.BrcFmri.R
\name{plotTimeSeries}
\alias{plotTimeSeries}
\title{Generic Plotting for Time Series}
\usage{
plotTimeSeries(x, ...)
}
\arguments{
\item{x}{the S3 object}

\item{...}{Arguments to be passed to methods, such as graphical parameters.}
}
\description{
Generic Plotting for Time Series
}

